### 2021-07-30 #3553 @whedon
[PRE REVIEW]: R-Opitools – An Opinion Analytical Tool for Big Digital Text Document (DTD)
**Submitting author:** @MAnalytics (<a href="http://orcid.org/0000-0002-9006-4934">Monsuru Adepeju</a>)
**Repository:** <a href="https://github.com/MAnalytics/opitools" target ="_blank">https://github.com/MAnalytics/opitools</a>
**Version:** v1.8.0
