### 2021-07-30 #3553 @kyleniemeyer
Hello @MAnalytics, thanks for your interest in JOSS. I noticed that this is a resubmission of #3083 (cc @kthyng). Can you please summarize the changes you have made in response to the original feedback?
