### 2021-07-30 #3553 @MAnalytics
> Hello @MAnalytics, thanks for your interest in JOSS. I noticed that this is a resubmission of #3083 (cc @kthyng). Can you please summarize the changes you have made in response to the original feedback?

Hello @kyleniemeyer @kthyng ,

The major criticism of the original submission was that the code was too narrow in scope; that it was too specific to one example application, and may be difficult for uptake in other areas. Consequently, I made the following changes:

#1. Modularizing the function 'opi_impact.R' by extracting (or separating) a new function 'word_imp.R' that will allow a text document (on any subject) to be explored for new themes from within the document.

#2. Provided two additional application examples of the codes in broader application areas (i.e. marketing & politics) to demonstrate the generalization of the new submission (please see the 'README.md' & the Vignette).

#3. The text of the paper and (many of) the documentation of the codes have been re-written in accordance to the changes made.

I hope this explanation answers your question.

Thank you very much,

Monsuru
