### 2021-07-30 #3553 @danielskatz
As another JOSS AEiC, I don't think this is enough of a change to justify a new paper under JOSS's [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion, but let's make this a formal discussion by the editors.  You will hear back in a week or so.
