### 2021-07-30 #3553 @MAnalytics
Dear @danielskatz @kyleniemeyer @kthyng ,

I greatly appreciate your suggestion.

Based on my understanding, the original submission was deemed substantial, but (in fact) the only critic was that it seems too specific to one subject area (law enforcements) upon which the method (in a journal article) was originally published (https://www.scirp.org/journal/paperinformation.aspx?paperid=107836). However, I warmly welcome the formal discussion against the criterion.

Thanks as I anticipate your favourable feedback.
M.A.
