### 2021-07-30 #3553 @danielskatz
I think it would be helpful if the paper discussed some of the use cases at a high level - basically, I think the paper should give some examples of who would use the software and for what, as part of the statement of need section.
