+++ paper.md
@@ -19,7 +19,7 @@ authors:
 
 # Statement of Need
 
-The recent proliferation of digital textual documents (DTD) across a wide range of application domains has not only heralded enormous data opportunities, but also new advances in the opinion mining of natural language texts [@Liu:2012; @Tsapa:2019]. However, in spite of the availability of wide variety of opinion mining tools [@Tai:2015; @Kiomourtzis:2014; @Wawer:2016; @Munson:2019], there has been a lack of mechanism for assessing the extent to which a selected theme or subject from within a DTD may have impacted the overall opinion expressed by the document. The `Opitools` package is designed in R language to address the aforementioned problem. In other words, given a digital text document, the package allows a user to test the statistical significance of the impacts produced by a selected theme or subject from the document on the overall opinion expressed by the document. The package can be applied to any DTD as long as the text records represent individual-level comments, such as posts (as on Twitter and Facebook) or comments (as in an online product reviews). The utility of `opitools` for social research is demonstrated in [@Adepeju:2021] for law enforcement application. Other applications, particularly in marketing and politics, are presented in the documentation.
+Since the year 2000, various computational intelligence techniques have been developed for analyzing sentiments of users either in the field of natural language processing or computational linguistics. To date, the majority of the techniques as deployed across various fields, such as Social research [@Somasundaran:2010; @Morency:2011; @Nikolovska:2020], Market research [@Feldman:2011; @Otaibi:2018], and in Electoral politics [@Ansari:2020; @Connor:2010] have focused largely on detecting subjectivity, and/or extracting and classifying sentiments and opinions in a text document. Building on this existing work, the current paper advances an opinion impact analytical tool, namely; `Opitools`, that does not only extracts inherent themes from within a digital text document (DTD), but also evaluates the extent to which a specified theme may have contributed to the overall opinion expressed in the document. Given the stated advancement, `Opitools` holds prospect for wider applications as compared to the existing techniques in the aforementioned application fields. For example, in Law enforcement, the package can be deployed to understand factors that drive public perception of the police service [@Adepeju:2021]; in Market research, to identify factors that underlie customers satisfaction in a product; and, in Electoral politics, to identify factors that may be influencing electoral opinions in an upcoming election. These examples are also demonstrated in documentation of the package (please see the `README.md` and the package `Vignette` for details). Lastly, it is argued that this contribution provides a framework by which some of the existing opinion mining tools [@Tai:2015; @Kiomourtzis:2014; @Wawer:2016; @Munson:2019] may be further advanced for wider research impacts.
 
 # Implementation
 
