+++ paper.bib
@@ -82,3 +82,64 @@
   publisher = {Birkhauser, 3rd Ed.},
   year=2006
 }
+
+@InProceedings{Somasundaran:2010,
+    author = {Somasundaran, S. and Wiebe, J.},
+    title = {Recognizing stances in ideological online debates},
+    booktitle = {CAAGET '10: Proceedings of the NAACL HLT 2010 Workshop on Computational Approaches to Analysis and Generation of Emotion in Text},
+    month = Jun,
+    year = 2010,
+    publisher = {ACM Press},
+    editor = {David Kurlander and Marc Brown and Ramana Rao},
+    pages = {116-124}
+}
+
+@InProceedings{Morency:2011,
+    author = {Morency, L. and Mihalcea, R. and Doshi, P},
+    title = {Towards Multimodal Sentiment Analysis: Harvesting Opinions from the Web},
+    booktitle = {Proceedings of the 13th International Conference on Multimodal Interfaces},
+    year = 2011,
+    publisher = {Association for Computing Machinery},
+    pages = {169–176}
+}
+
+@article{Nikolovska:2020,
+  title={“Show this thread”: policing, disruption and mobilisation through Twitter. An analysis of UK law enforcement tweeting practices during the Covid-19 pandemic},
+  author={Nikolovska, M.and Johnson, S. and Ekblom, P.},
+  journal={Crime Science},
+  year=2020,
+  volume=9,
+  pages={20}
+}
+
+@article{Ansari:2020,
+  title={Analysis of Political Sentiment Orientations on Twitter},
+  author={Ansari, MZ. and Aziz, M. B. and Siddiqui, M. O. and Mehra, H. and Singh, K. P.},
+  journal={Procedia Computer Science},
+  year=2020,
+  volume=167,
+  pages={1821-1828}
+}
+
+@InProceedings{Connor:2010,
+    author = {O'Connor, B. and Balasubramanyan, R. and Routledge, B. R. and Smith, N. A.},
+    title = {From Tweets to Polls: Linking Text Sentiment to Public Opinion Time Series},
+    booktitle = {Proceedings of the Fourth International Conference on Weblogs and Social Media, ICWSM 2010},
+    year = 2010
+}
+
+@InProceedings{Feldman:2011,
+    author = {Feldman, R. and Rosenfeld, B. and Bar-Haim, R. and Fresko, M.},
+    title = {The Stock Sonar - Sentiment Analysis of Stocks Based on a Hybrid Approach},
+    booktitle = {Proceedings of the Twenty-Third Conference on Innovative Applications of Artificial Intelligence},
+    year = 2011
+}
+
+
+@article{Otaibi:2018,
+  title={Customer Satisfaction Measurement using Sentiment Analysis},
+  author={Al-Otaibi, S. and Alnassar, A. and Alshahrani, A. and Al-Mubarak, A. and et al.},
+  journal={International Journal of Advanced Computer Science and Applications},
+  year=2018,
+  volume=9
+}
