### 2021-08-01 #3553 @MAnalytics
> I think it would be helpful if the paper discussed some of the use cases at a high level - basically, I think the paper should give some examples of who would use the software and for what, as part of the statement of need section.

@danielskatz @kyleniemeyer Sorry for the delay in response. I have now modified the statement of need section to include application areas of the software and some examples. Thanks.
MA
