+++ paper.md
@@ -23,12 +23,12 @@ Since the year 2000, various computational intelligence techniques have been dev
 
 # Implementation
 
-Having extracted a set of thematic keywords from a digital text document, the goal is to computationally classify the sentiments expressed in each text record into positive, negative or a neutral sentiment, using a lexicon-based classification approach [@Nielsen:2011; @Adepeju:2021]. The resulting sentiment scores are combined in order to estimate the overall opinion score of the document. To assess the impacts of a selected theme (or a subject) on the estimated opinion score, we simply ask the question; *If expected opinion scores were generated under the null hypothesis, how likely would we be to find a score higher than the estimated score?*. The question is answered by employing a non-parametric randomization testing [@Fisher:1935; @Good:2006] strategy involving random re-assignment of sentiment labels of the original text document to derive the expectation distribution. By comparing the observed score with the expectation distribution, the statistical significance of impacts is derived.
+Having extracted a set of thematic keywords from a digital text document, the goal is to computationally classify the sentiments expressed in each text record into positive, negative or a neutral sentiment, using a lexicon-based classification approach [@Adepeju:2021; @Nielsen:2011]. The resulting sentiment scores are combined in order to estimate the overall opinion score of the document. To assess the impacts of a selected theme (or a subject) on the estimated opinion score, we simply ask the question; *If expected opinion scores were generated under the null hypothesis, how likely would we be to find a score higher than the estimated score?*. The question is answered by employing a non-parametric randomization testing strategy [@Fisher:1935; @Good:2006] which involves random re-assignment of sentiment labels of the original text document to derive the expectation distribution, which is then compared with the observed score to obtain the statistical significance of the impacts.
 
 
 # Key Functionalities
 
-The `Opitools` package includes a function, i.e. `opi_imp`,  for exploring themes of a digital text document. To conduct the impact analysis, the main function is the `opi_impact()` which draws from two supporting functions, `opi_score()` and `opi_sim()`, to compute the observed opinion score and its expectations, respectively. Four different opinion score functions are embedded in the package, plus a `'fun'` parameter to allow a user to specify their own pre-defined score functions. This provision is to further facilitate the uptake of the `Opitools` in any application field.
+The package contains document exploratory functions, including `opi_imp`,  for extracting themes from a digital text document. To conduct the impact analysis, the `opi_impact()` function draws from two supporting functions, `opi_score()` and `opi_sim()`, to compute the observed opinion score and its expectations, respectively. Different types of opinion score functions are embedded in the package for selection based on the research question been investigated. A `'fun'` parameter is also included to allow the integration of pre-defined user score functions. This feature is to further facilitate the uptake of the package in more application fields.
 
 # Acknowledgment
 
