### 2021-08-05 #3553 @danielskatz
(based on previous discussion and slack comments)
