+++ vignettes/opitools-vignette.Rmd
@@ -399,7 +399,7 @@ col2 <- c("Does COVID-19 pandemic influence public opinion on neighourhood polic
 col3 <- c("`policing_dtd`","`reviews_dtd`", "`reviews_dtd`", "`debate_dtd`")
 col4 <- c("`covid_theme`","`refreshment_theme`", "`signage_theme`", "direct input")
 col5 <- c("two.sided", "two.sided", "two.sided", "two.sided")
-col6 <- c("(P - N)/(P + N)*100", "(P - N)/(P + N)*100", "(P - N)/(P + N)*100", "(P - N)/(P + N)*100")
+col6 <- c("Polarity score", "Polarity score", "Polarity score", "Polarity score")
 col7 <- c("-5.88", "67.92", "67.92", "-0.33")
 col8 <- c("0.52", "0.01", "0.1", "0.93")
 tble4 <- data.frame(col1, col2, col3, col4, col5, col6, col7, col8)
