### 2021-08-11 #3605 @whedon
[REVIEW]: R-Opitools – An Opinion Analytical Tool for Big Digital Text Document (DTD)
**Submitting author:** @MAnalytics (<a href="http://orcid.org/0000-0002-9006-4934">Monsuru Adepeju</a>)
**Repository:** <a href="https://github.com/MAnalytics/opitools" target ="_blank">https://github.com/MAnalytics/opitools</a>
**Version:** v2.0.0
