### 2021-08-12 MAnalytics/opitools#3 @justinchuntingho
Automated Tests
With regard to the following requirement:
- Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I see you have created `testthat` tests for various functions, would you consider using a continuous integration platform (eg Github Actions) so that the tests will run automatically after a push or pull request?

[Here](https://www.rostrum.blog/2020/08/09/ghactions-pkgs/) is a blog on how to set up Github Actions using `usethis` and [here](https://r-pkgs.org/r-cmd-check.html) is some instructions on using `Travis-CI`.

Part of JOSS review: openjournals/joss-reviews#3605
