+++ cran-comments.md
@@ -1,8 +1,0 @@
-## Test environments
-* local OS Windows, R 4.0.2
-* Debian Linux (R-devel)
-* win-builder (devel and release)
-
-
-## R CMD check results
-There were no ERRORs, WARNINGs or NOTEs. 
