+++ tests/testthat.R
@@ -2,3 +2,4 @@ library(testthat)
 library(opitools)
 
 test_check("opitools")
+#edit
