+++ tests/testthat.R
@@ -2,4 +2,4 @@ library(testthat)
 library(opitools)
 
 test_check("opitools")
-#edit
+
