+++ DESCRIPTION
@@ -45,6 +45,5 @@ Suggests:
   rmarkdown,
   testthat,
   rvest,
-  kableExtra,
-  markdown
+  kableExtra
 VignetteBuilder: knitr
