+++ tests/testthat.R
@@ -3,4 +3,4 @@ library(opitools)
 
 test_check("opitools")
 
-#test
+
