### 2021-08-12 MAnalytics/opitools#3 @MAnalytics
@justinchuntingho Thank you very much for the recommendation. I have now implemented the Github Actions) so that the tests will run automatically after a push or pull request
