### 2021-08-17 fc1e39c
Remove README header Add test to .gitignore
