### 2021-08-18 5ad903a
Remove ref. to pplot in code and README for opi_impact
