+++ README.md
@@ -51,8 +51,7 @@ The impact analysis can be conducted as follows:
 #Running the analysis
 
 results <- opi_impact(textdoc = policing_dtd, theme_keys=covid_theme, metric = 1,
-                       fun = NULL, nsim = 99, alternative="two.sided",
-                       pplot = TRUE, quiet=FALSE)
+                       fun = NULL, nsim = 99, alternative="two.sided", quiet=FALSE)
                        
 print(results)
 
