### 2021-08-24 MAnalytics/opitools#4 @jaeyk
paper.md wording 
I recommend tidying up some wordings in the `paper.md`. The following are examples I think that you can improve upon.

"either in the field of natural language processing or computational linguistics" -> "natural language processing or computational linguistics" are not mutually exclusive categories.

"including social sciences [@Somasundaran:2010; @Morency:2011; @Nikolovska:2020], market research [@Feldman:2011; @Otaibi:2018], and electoral politics [@Ansari:2020; @Connor:2010]" -> "electoral politics is a domain in social sciences, so listing these words as equal concepts would be not most appropriate."
