### 2021-08-24 #3605 @jaeyk
https://github.com/MAnalytics/opitools/issues/4 @MAnalytics Excellent R pkg. I recommend tidying up some wordings in the `paper.md`.
