+++ vignettes/opitools-vignette.Rmd
@@ -236,19 +236,19 @@ knitr::include_graphics("zipf.png")
 
 For a natural language text, the Zipf's distribution plot has a negative slope with each term falling on a straight line. Any variance from this (ideal) trend can be attributed to imperfections in the term usage. For example, the presence of strange terms or 'made-up' words can cause a deviation from the ideal trend. The result of our example dataaset is shown in Figure `r fig$ref("figs1")`. The graph can be divided into the three sections: the upper, the middle and the lower sections. By fitting a regression line (an ideal Zipf's distribution), we can see what the slope of the upper section is quite different from the middle and the lower sections of the graph. The variance at the high rank terms (usually, common terms, such as 'the', 'of', and 'at',) indicates an imperfection because of lack of perfect alignment with the straight line. For Twitter data, this variance suggests a significant use of abbreviated terms, such as using "&" or "nd" instead of the term "`and`". Apart from the small variance at the upper section of the graph, we can state that the law holds within most parts of the document.
 
-The second exploratory function `'word_imp'` can be used to highlight the importance of each terms in a DTD. The level of importance can then be used to identify different themes (or subjects) from the DTD. Two measures, namely (i) `'term frequency (tf)'` and `term frequency inverse document frequency (tf-idf)` [@Silge2016] can be used to highlight the importance of terms in a DTD. In Figure 2, results 1, 2 and 3 are generated from the datasets; `policing_dtd`, `reviews_dtd` and `debate_dtd`, respectively. The results of the measures `'tf'` and `'tf_idf'` for each data sample are labeled as A and B, respectively. The function draws from `wordcloud2` package in R [@Dawei2018], in order to highlight words importance as shown in Figure 2.
+The second exploratory function `'word_imp'` can be used to highlight the importance of each terms in a DTD. The level of importance can then be used to identify different themes (or subjects) from the DTD. Two measures, namely (i) `'term frequency (tf)'` and `term frequency inverse document frequency (tf-idf)` [@Silge2016] can be used to highlight the importance of terms in a DTD. In Figure 2, results 1, 2 and 3 are generated from the datasets; `policing_dtd`, `debate_dtd` `reviews_dtd`, respectively. The results for each data sample based on the `'tf'` and `'tf_idf'` measures are labeled as A and B, respectively. The function draws from `wordcloud2` package in R [@Dawei2018], in order to highlight words importance as shown in Figure 2.
 
 ```{r, message=FALSE, include = TRUE, eval=FALSE}
 
 #Load datasets
 
 data("policing_dtd")
-data("reviews_dtd")
 data("debate_dtd")
+data("reviews_dtd")
 
 
 #Words importance of public tweets on neighbourhood policing 
-#based on  (a) ‘tf’ (b) 'tf-idf'  
+#based on  (a) ‘tf’ and (b) 'tf-idf' metrics
 p1a <- word_imp(textdoc = policing_dtd, metric= "tf", 
                            words_to_filter=c("police","policing"))
 
@@ -258,21 +258,21 @@ p1b <- word_imp(textdoc = policing_dtd, metric= "tf-idf",
 #Note: 'words_to_filter' parameter is used to eliminate non-necessary words that 
 #may be too dominant in the DTD.
 
-#Words importance of customer reviews of a transport service 
-#based on (a) ‘tf’ (b) 'tf-idf' 
-p2a <- word_imp(textdoc = reviews_dtd, metric= "tf", 
-                words_to_filter=c("station")) 
-
-p2b <- word_imp(textdoc = reviews_dtd, metric= "tf-idf",  
-                words_to_filter=c("station")) 
-
 #Words importance  of comments on a video of a political debate 
-#based on (a) ‘tf’ (b) 'tf-idf'
-p3a <- word_imp(textdoc = debate_dtd, metric= "tf", 
[…]
