### 2021-08-24 50d750b
Update issue templates
added another template
