+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,20 @@
+---
+name: Feature request
+about: Kindly suggest an idea (an argument, a function, etc.) for this project
+title: "[Add feature]:"
+labels: documentation
+assignees: MAnalytics
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+
+**Are alternatives you've considered? Please briefly describe**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional information**
+Add any other information or screenshots about the feature request here.
