### 2021-08-24 MAnalytics/opitools#4 @MAnalytics
@jaeyk The wordings (as indicated) have now been tidied.
