### 2021-08-24 #3605 @jaeyk
@MAnalytics As for the installation guideline, it might be useful to explain that the {remotes} is an extra package that users need to install.
