+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: 2.7.3
+pkgdown: 1.6.1
+pkgdown_sha: ~
+articles:
+  opitools-vignette: opitools-vignette.html
+last_built: 2021-08-24T23:33Z
+
