+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   opitools-vignette: opitools-vignette.html
-last_built: 2021-08-25T07:13Z
+last_built: 2021-08-25T07:24Z
 
