+++ README.md
@@ -10,6 +10,8 @@ An R-package for analyzing Opinions in Big Digital Text Document (DTD)
 
 The `opitools` is an R package for exploring a digital text document (DTD) as well as performing the impact analysis of the opinions expressed by the DTD. The package is particularly suited for opinion-based DTD, such as individual-level posts (e.g. commentaries on Twitter or Facebook) and comments (as in an online product reviews). First, an `exploratory` function `word_imp` can be used to identify words relating to different themes (or subjects) that are referenced in a DTD. The `opi_impact` function can then be utilized to investigate whether an identified/specified theme (or subject) impacts the overall opinion expressed by the document. The potentials of `opitools` for application across a wide range of domains is described briefly here (see the `vignette` for details). 
 
+[Click here](https://manalytics.github.io/opitools/index.html) to visit the built website for the package.
+
 ## Installation from `CRAN`
 
 From an R console, type:
