### 2021-08-25 #3605 @MAnalytics
Dear @jaeyk @justinchuntingho @cjbarrie @ajstewartlang

All of the above have now been addressed accordingly. The package website can be accessed here: https://manalytics.github.io/opitools/index.html

Thanks,
MA.
