### 2021-08-25 c71d56d
Fix documentation typos; edit paper.md
