### 2021-08-25 MAnalytics/opitools#4 @cjbarrie
Sorry to reopen, but didn't want to create multiple Issues. I have a few further language editing suggestions:

- "that does not only extracts inherent themes" should be "that not only extracts inherent themes'
- "holds prospect for wider applications" reads a bit oddly, how about just "has wider applications"
- then I'd delete "compared to the existing techniques, in the aforementioned application fields" as this is pretty much implicit anyway
- I would then also delete the lines "The demonstration of these examples can be found in the documentation of the package (please see the `README.md` file and the package `Vignette` for details). Lastly, it is argued that this contribution provides a framework by which some of the existing opinion mining tools [@Tai:2015; @Kiomourtzis:2014; @Wawer:2016; @Munson:2019] may be further advanced for wider research impacts." as I'm not sure these really add much without further comment on why the package advances these existing tools. Given this section is really supposed to be just a short Statement of Need, deleting these lines is appropriate.
- In the section "Key Functionalities" I would remove reference to function names (which may change) and instead just report the high-level functionality
