### 2021-08-25 #3605 @cjbarrie
I have created a PR, which just tweaks a few things. Otherwise, my editing suggestions are limited to those edits to paper.md described in https://github.com/MAnalytics/opitools/issues/4.

Once these have been resolved, I'd be happy to accept!

Great work @MAnalytics !
