### 2021-08-25 MAnalytics/opitools#4 @MAnalytics
@cjbarrie Thank you very much for the excellent reviews. All the changes have been made accordingly.

MA.
