### 2021-08-31 #3605 @ajstewartlang
Hi @MAnalytics the two suggested DOIs above look correct to me - can you double check and add them to the paper please? Thanks :-)
