+++ paper.bib
@@ -107,6 +107,7 @@
   title={“Show this thread”: policing, disruption and mobilisation through Twitter. An analysis of UK law enforcement tweeting practices during the Covid-19 pandemic},
   author={Nikolovska, M.and Johnson, S. and Ekblom, P.},
   journal={Crime Science},
+  doi={10.1186/s40163-020-00129-2},
   year=2020,
   volume=9,
   pages={20}
@@ -116,6 +117,7 @@
   title={Analysis of Political Sentiment Orientations on Twitter},
   author={Ansari, MZ. and Aziz, M. B. and Siddiqui, M. O. and Mehra, H. and Singh, K. P.},
   journal={Procedia Computer Science},
+  doi={10.1016/j.procs.2020.03.201},
   year=2020,
   volume=167,
   pages={1821-1828}
