### 2021-08-31 #3605 @ajstewartlang
Many thanks @cjbarrie @jaeyk @justinchuntingho for your helpful and detailed reviews. I've had a read through myself @MAnalytics and agree that this is a very nice and useful piece of software. Congrats!

@MAnalytics can I now ask you to do the following please?

    Make a tagged release of the software.
    Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
    Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add your ORCID.
    Please list the DOI of the archived version here.
